+++ era-compiler-solidity/tests/cli/force_evmla.rs
@@ -100,3 +100,21 @@ fn with_force_evmla_standard_json_mode(target: Target) -> anyhow::Result<()> {
 
     Ok(())
 }
+
+#[test_case(Target::EraVM)]
+#[test_case(Target::EVM)]
+fn standard_json_deprecated(target: Target) -> anyhow::Result<()> {
+    common::setup()?;
+
+    let args = &[
+        "--standard-json",
+        common::TEST_SOLIDITY_STANDARD_JSON_ZKSOLC_FORCE_EVMLA,
+    ];
+
+    let result = cli::execute_zksolc_with_target(args, target)?;
+    result.success().stdout(predicate::str::contains(
+        "The `forceEVMLA` setting is deprecated. Please use `codegen: 'evmla'` instead.",
+    ));
+
+    Ok(())
+}